requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
Django==5.2.4
django-allauth==65.10.0
django-crispy-forms==2.4
gunicorn==23.0.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
sqlparse==0.5.3
whitenoise==6.9.0
